# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fornax",
    "description": "",
    "contact": {
      "name": "Evorsio"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local"
    },
    {
      "url": "https://fornax.com",
      "description": "Production"
    }
  ],
  "paths": {
    "/api/": {
      "get": {
        "tags": [
          "General"
        ],
        "operationId": "root",
        "responses": {
          "200": {
            "description": "服务正常"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResp"
                }
              }
            }
          },
          "400": {
            "description": "邮箱或验证码不能为空"
          },
          "401": {
            "description": "验证码错误或已过期"
          },
          "500": {
            "description": "服务器内部错误"
          }
        }
      }
    },
    "/api/auth/send-code": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "send_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCodeReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "验证码发送成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendCodeResp"
                }
              }
            }
          },
          "400": {
            "description": "邮箱不能为空"
          },
          "500": {
            "description": "服务器内部错误"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LoginReq": {
        "type": "object",
        "required": [
          "email",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "LoginResp": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "SendCodeReq": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "SendCodeResp": {
        "type": "object",
        "required": [
          "msg"
        ],
        "properties": {
          "msg": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "认证接口"
    }
  ]
}